Restructure the tree
add_host is used to add containers to the inentory dynamically so
that we can treat the container and non-container novajoins in the
same way, rather than writing yucky container specific code.

add_host always runs on the controller host, but it selects a host
seemingly at random to be the host it "runs" on, and then bypasses
the rest of the host loop.  If you're lucky, the host it chooses is
the undercloud, and all goes well.  If you're unlucky (more likely
than not), the host it chooses is not the undercloud, and because of
when: conditions, the task is skipped.

The best way to deal with this to break things up into separate roles.
This is what this patch currently does.  We now have three roles -
tls-everywhere-helper-common, tls-everywhere-helper-overcloud and
tls-everywhere-helper-undercloud.

A checkin to remove the old code will follow this one.